Hide redundant entities by default
The AnalogOutput cluster is used to write target position to the device.
The present_value reported/read provides the current position/starting position when the device is in motion.

Both of these functions are re-mapped into the ZCL WindowCovering cluster, so the number entity created is redundant.

This also pre-emptively hides the MultistateOutput cluster, though there is currently no cluster handler implemented for this.

Apply pre-commit auto fixes